# remove search index generation from dev build