Part of https://github.com/lit/lit.dev/issues/1201

### Context

While profiling lit.dev, I found this as an easy performance win to make dev mode builds faster.

In the old days, the Lit.dev search was done locally, so the search index could be used and debugged locally. However now the search index is instead deployed to Algolia. Thus the several seconds it takes to create the search index can be skipped when building the site in dev mode.

### Results

Before this change, tested with `dev:build`:

`[11ty] Wrote 373 files in 15.58 seconds (41.8ms each, v2.0.1)`

After this change:

`[11ty] Wrote 373 files in 12.15 seconds (32.6ms each, v2.0.1)`